Refactor Autonity Contract params
