+++ core/genesis.go
@@ -56,7 +56,7 @@ type Genesis struct {
 	Coinbase               common.Address      `json:"coinbase"`
 	Alloc                  GenesisAlloc        `json:"alloc"      gencodec:"required"`
 	Validators             []string            `json:"validators"`
-	AutonityContractConfig *AutonityContract   `json:"autonityContract"`
+
 
 	// These fields are used for consensus tests. Please don't use them
 	// in actual genesis blocks.
