+++ params/autonity_contract_params.go
@@ -1,4 +1,4 @@
-package core
+package params
 
 import (
 	"errors"
